+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/component/ViewSurface.kt
@@ -5,7 +5,10 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
+import site.remlit.snowdrop.util.LocalContentWarningController
+import site.remlit.snowdrop.util.rememberContentWarningController
 
 /**
  * Basis of every View. Sets a standard background and fills the entire screen with its content.
@@ -13,8 +16,13 @@ import androidx.compose.ui.Modifier
  * */
 @Composable
 fun ViewSurface(content: @Composable () -> Unit) {
-	Column(
-		modifier = Modifier.background(MaterialTheme.colorScheme.background)
-			.fillMaxSize()
-	) { content() }
+	val contentWarningController = rememberContentWarningController()
+	CompositionLocalProvider(LocalContentWarningController provides contentWarningController) {
+		Column(
+			modifier = Modifier.background(MaterialTheme.colorScheme.background)
+				.fillMaxSize()
+		) {
+			content()
+		}
+	}
 }