+++ CHANGELOG.md
@@ -7,11 +7,13 @@
 - List timelines
 - Hide media viewer decorations when you zoom in
 - Emoji tooltips with their names
+- Thread view has a button to toggle all content warnings for all posts in a thread
 
 ### Fixed
 - Timelines not updating properly
 - Emojis missing from places where they belong
 - Weird padding when having content warning field in focus in compose page
+- Content warning state persists when posts go off-screen
 
 ## [0.0.5-alpha]